requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
requests
python-dotenv
werkzeug
Flask-Migrate
psycopg
